Make a few changes to the shared data containers in keep-core/relay.
In particular, we use uints for group and groupCount, and introduce GroupID,
which is the id of the current node within the relay group. We also make Value
and Timestamp public in Entry, and make Value a fixed-length byte array.

This preps us a little more for actually interacting with BLS, though none of
the types are completely final.