Flesh out share validation skeleton when generating a relay entry.
We now generate a fake BLS signature share for the current node, "broadcast"
it, then go into verification mode. We declare a callback that will stand in
for share broadcasting, and we then go into verification mode, where we listen
to incoming shares on a "passed in" channel. We call a stub verification for
each share, panic on an invalid share as a placeholder, and once the threshold
number of verified shares are received, go into submission mode and generate a
relay entry that we "submit" to the chain.