Break member struct into structs for various phases
The one Member struct becomes a LocalMember, SharingMember, JustifyingMember,
and Member struct. Each one of these represents a phase in DKG, except
SharingMember which represents both the commitment and private share
receiving/validation phase. These can occur at the same time without any
issues, so we combine the struct.

Each struct has an associated method to transition it to the next phase, in
which we do any final work (e.g., when we go from LocalMember to SharingMember
we compute the private member shares that are to be sent to each member of the
threshold group). Each struct also exposes the various methods that make sense
for its phase.

The Member struct now represents a fully-fleshed-out member, ready to
participate in threshold signatures.

Note that there are still a few todos around properly tracking
accusation/justification. There is also some validation that needs to happen so
that we don't generate a final Member if we've failed to meet our threshold
requirements for qualified members.