Add DKG package and procedure
DKG.Execute takes a block counter, broadcast channel, group size, and
threshold, and attempts to initialize a single member in a threshold group.
Multiple DKG.Execute calls executed in parallel, combined with a local
broadcast channel, will result in a local DKG process occurring.

This gives us the infrastructure to do a local non-networked test of DKG, with
interface stubs to interact with a networked version as well.

Note that currently the DKG relies entirely on block timeouts for phase
synchronization. We'll need to consider whether that's necessary throughout.