Add signature share generation and verification
thresholdgroup.Member now has SignatureShare(message) to compute its share of a
threshold signature, and VerifySignature(shares, message) to take a set of
shares and verify that they are a valid signature of the message for this
group.

The shares are all passed around as byte arrays, so that the details of BLS are
tucked away behind everything with the exception of BLS IDs (which we may hide
later).