Add vendor directory to github.
Adding the vendor directory to our git repo is recommended by the godep
authors. The build-keep-client-docker-img.sh script works on the
assumption that the vendor directory exists.  Otherwise, the build
server would need to run dep ensure, which is not preferrable. For more
details on why the vendor directory should be included, see:
https://github.com/tools/godep/blob/master/Readme.md