Fix docker build, but really this time though
We re-separate the dep ensure step so that it doesn't run when a simple Go file
changes. We also fix the directory paths so that imports that reference the
full keep-core GitHub path will resolve correctly (currently no files with
these references are compiled, which is how we missed it, but they're coming).