relay: introduce protobuf types that we can use to
communicate across service boundaries and even within a service.

Specfically in regards to a within a service, we can iterate on expected
types. This is in addition to maintaining a spec.

For the future, this will let us keep tasks that are expensive in their
own packages. This will allow our users to run the relay client in a
cluster, where chain events are their own backend, syncing events and
streaming them to our relay node. DKG/BLS can be run on a machine and
block progression.