Initial outline for random beacon yellowish paper
The yellowish paper will probably become a complete spec with information on
dependent technologies, links to work that inspired the relay, and details on
wire formats. The goal would be to be able to write a compatible client from
scratch from the yellow paper, if we have the time/bandwidth to get that far.