Introduce interfaces and types for network.
Relay Clients require a clean interface to be able to join and
communicate with Groups over a panopoly of network protocols. We
abstract this away through the minimal number of fucntions and types
needed to abstract the concept of Groups and Peer-to-Peer networks.

This is a heavy work in progress.