Update miner censorship mitigation to be seed commit-reveal
Future blockhash isn't ideal since it allows the miner of the future block to
skip mining the block if they know the hash will select their group + their
group's signature of the previous entry would be bad for the miner.
Commit-reveal of the request seed makes the miner accept a request first, then
get the seed once the request is already queued up.