Add pkg/beacon/relay/dkg package
This package currently just contains the messages that are used for relay DKG.
These are very similar to existing ones in beacon.relay, but carry some
additional information that will be needed as we look to abstract the network
sender/receiver info from the DKG-specific BLS ids. They also carry most BLS
ids as pointers instead of values.