Add pkg/chain and rearrange interfaces
The new package contains abstract chain-interaction interfaces. For now, the
main handle to a chain provides access to an implementor of the beacon's
expected chain interface. The idea is to use this mechanism to provide, from
one chain handle, access to multiple underlying contract interfaces, with the
base functionality being specified by the consumer and the chain packages
bridging the consumer's API with the chain API.

This may end up needing further adjustment.