Correctly handle block-heigh-in-the-past for BlockWaiter
When the BlockWatier found itself notifying a block height before the
current one, it would immediately try to send a message to the new
waiter channel. However, since the waiter channel is unbuffered and has
not been returned to the consumer yet, that would block the function
indefinitely.

We now message the watier channel in a separate goroutine, which can
unblock as soon as the caller consumes the message.