Add a concept of net.ProtocolIdentifier
ProtocolIdentifier is meant to allow us to separate the identifier used
internally for a network layer, to identify a given node, from
identifiers needed for a given protocol, for example BLS IDs. We start
by introducing net.ProtocolIdentifier as a type, introducing a Message
interface for messages that carry both network and protocol sender ids,
and add a RegisterIdentifier method to the BroadcastChannel interface
that allows the protocol layer to map a network identifier to a protocol
identifier.

Consumers are not yet using this abstraction.