Adjust net/local to understand protocol identifiers
Local channels are reworked a bit so that each consumer will have its
own channel instance, exactly as if they were across a network, but the
instances with the same channel name will communicate with each other.
Each channel instance also has a random identifier. We introduce an
internal BasicMessage for easy instantiation of the Message interface
without external exposure.

Also clarify that the local package is not meant for real use.