Move local chain implementation to a handle-based model
Similar to what we'll be doing with the network layer, you can call
local.Connect() and get a handle that exposes the chain interface. We
move the stub implementation of the interface to a backing struct that
allows for the addition of submitting group public keys in local
operation.

Another side-effect is we make the local block counter implementation
accessible only through a local chain.Handle.

Overall, this should make interactions with the local chain
implementation much closer to interactions with a real one.