Rename net.ClientIdentifier to net.TransportIdentifier
Also tweak various places where we're referring to a netID or networkID
or networkIdentifier to reflect the new term. This term is both clearer
(it's really an identifier for whatever transport we're using for our
messages) and gives us some consistency across different places where
we're using it in terms of naming.