Expose method to connect to libp2p, recv a Provider
External consumers of net will want a mechanism to connect to a peer to
peer network. Our Connect() takes a context and a port (unfortunately a
requirement for now) and returns a `net.Provider`. The `net.Provider` is
used by external consumers in interacting with the net package.

Includes unexporting channel manager functions.