Rework DKG in terms of keyGenerationState interface
There is currently only one such state, which immediately ends the state
machine. However, this sets the groundwork for adding the remaining
states back.

The state machine itself enforces an active block period. If a state
does not complete within that period, the key generation process fails
and an error is returned. When a state is final, indicated by returning
a nil next state, it is expected to have a valid thresholdgroup.Member.
If it does not, the key generation process fails and an error is
returned.

State transitions are expected to occur as a result of a network
message.

Finally, incoming messages are serialized on an unbuffered channel so
states don't have to manage synchronization overhead.