Reorganization to fix go circular dependency
This also is a much better what of doing the configuration.

Put the top-level Config struct and the routines that read it from
the filesystem in the top-level keep-core package.

That struct then references substructs from the packages that need
config (e.g., ethereum.Config, later on libp2p.Config, etc),
and these are the objects that are passed to Connect or Init
methods for those packages.

A `TODO` comment has been left in the top level config.go file
where you will need to add the libp2p stuff.