FinalizeMember errors on qualified < threshold
FinalizeMember would return a member irrespective of how many qualified
members were intact at the end of DKG. In fact, DKG can be considered a
failure if fewer than `threshold` members are qualified at the end of
the process. FinalizeMember now returns an error in this case, and a
final Member otherwise.

To make this work correctly, the DKG states and state machine had to be
updated to deal with a state transition that results in an error.