Factor core data into thresholdgroup.memberCore
memberCore includes the data that's common to both distributed key
generation and threshold signing. Before, JustifyingMember was embedded
in the Member struct, but that pulled in a lot of data that the final
Member didn't need for threshold signing. We now factor out the common
data so that Member can jettison the key-generation-specific items.