Properly track invalid shares from other members
Previously, an invalid share would not leave a trace in receivedShares.
receivedShares now associates with a pointer to a secret key instead of
the raw value, so we now insert a nil pointer for shares that were
received but were invalid. This allows SharesComplete to always be an
accurate reflection of whether all members have *sent* shares, even if
they sent *invalid* shares.