Optimize locks, introduce pubsub lock to type
Rather than taking a huge lock for the duration of a function, we take the
lock only when accessing the map and avoid holding it for operations
that don't pertain to accessing the map of channels.

We also introduce a lock for the pubsub state, as it is a shared
resources and we may access it concurrently.