Avoid mutating self share during group sk gen
When SecretShareForID started returning a pointer, using `.Add` on the
member's own secret share led to updating the secret share in place
rather than generating a new group secret key. This led to broken
signatures. We now copy the member's own secret share into a new
container and add to that container.