Fixed polynomial degree and threshold description
According to Shamir the polynomial degree should equal `k-1`, then `k` members of the group are sufficient to reconstruct the secret.

In DKG and our implementation, the `threshold` is a maximum number of dishonest group members when the protocol is still secure, this implies the degree of polynomial to be equal to `threshold`.

In such polynomial number of polynomial coefficients equals `threshold + 1`, which is the same as a number of `secretShares` and a number of `shareCommitments` for each member.