Identity, Peerstore, and Pubsub state mods
Modify protobuf Identity type to only carry the public key, change our
marshalling and unmarshalling code accordingly. To differentiate between
incoming Identities and the client's identity, we rename the identity
state on the channel to `clientIdentity`. This helps disambiguate
between the identity local to us and that of an external party.