Modify Recv and proceessMessage
Recv previously buffered messages expecting the caller to continuously
call Recv. This wasn't the interaction model: instead, the caller will
register a callback, we'll add it to a list of callbacks, and on an
incoming message, spin up a goroutine to call all handlers on a given
message.

Furthermore, we simplify the return of `deliver` as well as rename
`unmarshaled` to `unmarshaler`.