Add message buffering
There may be a window of time where our client begins accepting messages
on the network before the Protocol has initiated a callback. In this
case, we make the special exception of buffering message right up until
we get our first handler. We immediately drain the buffer and then
disable it for further use.