Remove timer in busy loop, handle formatting
We remove the unnecessary ticker from our busy loop: Next() naturally
does this by being a blocking call. As messages are available, we'll
process them as fast as they are available. In addition, Next() handles
checking the state of the context, so we remove that check as well.

Also, fix formatting on the BasicMessage assignment.