High-level partial restructure of beacon kickoff
Right now the beacon is set up to take the appropriate interfaces and
then run a single distributed key generation. This isn't the final
version of the code, but gets the code to a place where, once chain and
network integration is set up, it can be piped in and a DKG can
successfully occur.