Restructure relay packages to avoid import cycles
As more code gets filled in, there's a strong possibility for import
cycles if chain and config pieces aren't isolated. There is now a
relay.config package and a relay.chain package. These are very
lightweight packages, with relay.chain.Interface replacing
relay.ChainInterface, and relay.config.Chain replacing relay.Config.

This also allows for a clear distinction between on-chain configuration
and config-file-bsaed configuration, which will likely prove useful.

References outside of the relay package are to relayconfig.Chain and
relaychain.Interface to ensure that the provenance is clear.