Flesh out relaychain interface
Introduce SubmitRelayEntryCandidate to the interface. This allows us to
glue together Submitting a relay entry candidate to the chain for
consideration. If the entry (relay response) is accepted as a valid
entry within the appropriate number of blocks, then we'll proceed
forward (distributing rewards, and moving on to selecting the next
group).