Add first draft of DKG protocol document
Group selection and distributed key generation requires a protocol with
sufficient resilience against misbehavior. This variant of GJKR aborts only if
over 25% of participants are inactive or misbehaving, and honest-and-available
participants will agree on who tried to share invalid values.

There is also a well-defined ordering and incentive structure for submitting
group creation results on-chain, and the honest majority assumption should
guarantee that conflicting claims of results are resolved correctly.