Track if Promise succeeded to execute the right callback
We had a bug in `Promise` that if we called `OnFailure` and the `Promise`
had already completed, the callback was called irrespective of whether
the promise had completed with a failure or a success. The same was
happening for `OnSuccess`.

In order to avoid it, in addition to `isComplete` bool, we now have
a `succeeded` boolean to determine if the given callback should be
invoked.