Take advantage of caching for some generated code
Rather than throw away all of the benefits we get from caching by moving
our directory COPY to the top, COPY and generate the generated
files we expect not to rely on imports in the project first. Then,
for generated code that does rely on external code, we generate
AFTER our directory copy.