Refactor tests to conform to existing style
We use existing table maps, which define a map with a test name ahead of time,
and run the individual cases as subtests so that the test output includes the
name automatically.

Also refactored file package name back to libp2p and unexported structure
and package names that were exported ... in an effort to prevent the cyclic
redundancy error that was caused by referencing the config.Config struct
that this commit removes.  So, now we have no cyclic errors and no
unnecessarily exported symbols.