Introduce relay.Node and relay.NewNode
relay.Node is a container for a single relay node's state. It has a
stake id and carries references to external interactors like the
net.Provider, chain.BlockCounter, and config.Chain that the node is
using. relay.NewNode can be used to create a new node state with
external interactors set, and beacon.go does just this.