Implement group joining in relay
Specifically, when a new entry is seen, the relay node checks if it
belongs to the stakers selected by that relay entry. If so, it initiates
distributed key generation, and, if that succeeds, submits the group
public key to the chain.

It also locally tracks the thresholdgroup.Member generated for a group
that the node attempts to join, so that when the group is successfully
registered the local node can have a reference to its private key share.