Removed g from hash input
In the original paper, generator point of elliptic curve is put into the
has function. However, in go-ethereum g point is a constant so there is
no need to include it in sum256 as it can only reduce randomness.

The other thing is that we were putting G=N+1 (Paillier parameter) into
hash instead of elliptic curve generator point and this is obviously
wrong.