Directly populate libp2p.Config from file
libp2p.Config was being manually populated from Bootstrap and Node
entries in the config file. These two are removed in favor of a direct
reference to libp2p.Config. Alongside that, the --bootstrap flag is
removed, as it was simply used to choose between entries in Bootstrap vs
Node when populating libp2p.Config.

This means only one entry exists in the config file for libp2p, so that
entries for Bootstrap.URLs become entries for LibP2P.Peers,
Bootstrap.Seed becomes LibP2P.Seed, and Node.Port becomes LibP2P.Port.

Existing checks ensure that if Seed is set and Peers is also set, we
will error out, since one of these is for the bootstrap node and the
other is for non-bootstrap nodes that want bootstrap nodes to connect
to.

Sample config file adjusted accordingly.