Keep secret value separately from the decommitment key
Although the previous approach with storing secret value along with a
decommitment key in one structure was correct, it was quite unwieldy in
a real world. Because how our workflow is structured, lack of
generics in go and difficult deserialization of elliptic curve points
from bytes, it's better for us to keep the secret value separately and
pass it later as a function argument for verification along with the
decommitment key.