Each registered instance of the callback is unique
As we may see many relay entries and relay requests, we need to mark
each saved instance of a callback as unique. A simple way to do this is
timestamping our entry. A practical instance of this helping is when we
unregister, we shouldn't unregister all active callbacks as some
goroutines may still be running.