Let to generate trapdoor commitment for multiple values
For the sake of T-ECDSA signature generation we need to be able to
generate a trapdoor commitment for multiple values. In this commit, I've
switched signature of the commit and verify function to allow for a
vararg secret argument.

All the passed secret messages are combined into one in order to
evaluate a commitment. I have also renamed `GenerateCommitment` function
to just `Generate` since it looks nicely with `commitment` package:
`commitment.GenerateCommitment` vs `commitment.Generate`.

I had to switch order of parameters in the `Verify` function as the
vararg param must be the last one.