Associate group public keys with request IDs
There was a strange concept of groupID being pushed through public key
submission that fell apart when it hit the chain because the chain
properly expected group public keys to be associated with the request ID
of the relay entry that triggered the group creation. This request ID is
now properly fed through the whole pipeline.