Only handle group public keys once per request ID
Each group registration event would trigger the appending of the public
key to the end of the list of public keys. This meant if one group
triggered multiple registration events for whatever reason, there would
be two entries in the group list, which would lead to bad member
selection later.

The request ID is now tracked as being seen or not seen, and if it was
seen the group public key for that request ID is not re-appended.