Existing identifiers shouldn't error
If we have a matching transport and protocol identifier pair, we
shouldn't error and halt dkg. Rather, if everything lines up, we should
continue. We should only error in the case where the transport
identifier and protocol identifier attempting to be registered don't
line up with what was previously registered.