Ensure that only one round of DKG runs at a time
For the purposes of our demo/M1, it's fine if we ignore multiple
requests for group formation. This masks an issue where a single request
can trigger multiple rounds of DKG over the same channel (which causes
an issue where we'll get conflicting protocol identifiers for a given
client over the same channel).