Rely on syncing our staking list
Rather than add our stake explicitly on success, we register our staker
added callback, and then attempt to sync our staking list in whole. We
make this call resilient to network timeouts and other issues by
retrying it every three seconds (as it's critical we have an up-to-date
staking list).