Get the T-ECDSA parameters naming right
1. We don't add `encrypted` prefix for values if there are no conflicts
with plaintext values. The type declaration is enough to say if the
given value is encrypted or not.

2. We use `*share` prefix everywhere we deal with shares except in ZKPs
- we don't want to complicate things there for now)

3. `k` is a signature factor, not signature multiple. Naming was updated
accordingly

4. We dropped `random` word from all parameter names. The fact something
is randomly picked is an implementation detail. Secret key factor is
still a secret key factor, signature factor is still a signature factor,
it doesn't matter if it was randomly picked or not.