Switch from EDDSA to ECDSA keys.
Swap out ed25519 (EDDSA) keys for secp256k1 (ECDSA) keys.

Given that ECDSA keys are faster to verify on chain, we want to drive
consistency by opting to have all of our keys be ECDSA (to avoid
ambiguity). We may revisit this decision, but this gives us a solid
starting point (though network identities may never make it on-chain).