Intruduced protocolParameters struct
This struct is intended to hold master trapdoor public key for peer as
well as hold all the ZKP parameters we'll generate individually for peer
in the future.

I have also modified the way how protocol parameters from other peers
are received to do not overwrite peer's own settings. That's just for a
better security.