Added function returning peer signer count for the group
With the new approach for common parameters setup, we'll be sending
unicast messages with commitments to all peer group members instead of
sending one broadcast message within a group. To make this happen, we
need to be able to figure out what's the number of peer signers in the
group.