Use predefined safe primes to construct ZKP test parameters
Since safe prime generation is expensive and we want our unit tests to
be blazing fast, we need to use a predefined safe prime values.

The other possible approach is to generate small safe primes, e.g.
128-bit size. However, since we want our unit tests to simulate a real
conditions as close as possible, the fixed 1024-bit primes are a better
solution.