Peer which tampered the message is blacklisted
Thoughts:

We can't have a broadcast-style complaint about tampered message since there is no way, a peer can prove the complaint is valid or not. There are only two parties involved in the complaint - sender who says the message was not altered and receiver saying the message was altered. Who should we trust? The receiver could tamper the message on its own to make a false accusation against the sender and sender could really tamper the message.

Because of that, we can't blacklist peer from the network in one move. Only the receiver can individually blacklist the sender.

Let's assume we do blacklisting and we have a network of ~50 peers at the beginning.
The message has been tampered so the sender is blacklisted for 10 minutes by the given receiver. Still, the malicious sender can interact with 49 other peers (which may even relay the message to the peer who blacklisted the malicious sender). In order to effectively elminiate the malicious sender from the network, we need to observe 49 tampered messages on 49 individual peers which will all blacklist the sender.

There are two scenarios possible:

1. We have an adversary that hacked peer's network but does not know peer's secret key.

    In this case, the only thing the adversary can do is to flood the network with tampered messages. Even without the blacklisting, when all the peers drop their connections with sender, adversary can no longer authenticate because it can't prove on-chain identity without knowing the secret key. We don't need blacklisting in this case.

2. We have a malicious peer or an adversary hacked peer entirely and knows the secret key.

   Even if we drop the connection, malicious peer that can prove an on-chain identity with a minimum stake can quickly connect again. The only way to prevent the flood of tampered messages is to blaclist the peer. The possible scenario is as follows: malicious peer tries to hack some Keep protocols. Since we use ZKPs and can prove malicious activity on-chain, adversary peer is slashed. Adversary gets angry and decides to break the communication in the network to make it unreliable. It starts tampering relayed messages. (We don't consider dropped messages here, it's an entirely different scenario). The tampered message is rejected by the next relay peer and malicious adversary gets disconnected. The message can be retransmitted without the adversary on the routing path. To make this work, we should blacklist adversary for at least the time that is needed for a message to be retransmitted. 

Realistically, adversary knowing the secret key will probably try more sophisticated attacks than just tampering relayed messages, but since there is a relatively easy way to eliminate this threat, I think we should blacklist for some set amount of time. We may decide to tweak it to some sort of exponential backoff-style setup, but in the first version, I'd just add an on-chain, global value for a blacklist period.