Add relay submit for testing the app
Though a beacond may be initialized, peers in our network listen for an
initial relay entry, which is the seed to the network, to kick off the
initial group selection. This initial group selection determines which
peers are selected to run DKG to create a group. Those that succeed can
then respond to relay requests.

All of this hinges on a valid seed relay entry. This cli command enables
a tester to submit a random value to simulate group selection.