Add a cli command to ping the network
Ping is a cli command that allows two users of the keep client to ensure
that a peer-to-peer mesh is constructed between participants without
building up all the state for the various protocols. The client exposes
a ping command which only tests the network interfaces, sends a ping
command over the wire (to a specific peer), and receives an unmarshalled
pong if everything is working correctly.