Change group signature to big int
Previously, group signature (entry.Entry.Value, relay.Entry) was a
32-byte fixed-sized array. Throughout the code, and in the abi, we
reference this value as a big int. Furthermore, we reference the other
values as a big int. To simplify the types, we change this value to a
big int, and enforce the 32-byte bound in the GenerateRelayEntry
function (via only taking the first 32 bytes).