Refactor and move code around for reuse
Move the methods that are reusable within this package (outside of the
context of an authenticated connection), into their own file
(network_handshake.go). This way, transport.go and
authenticated_connection.go (which explicitly implement libp2p
interfaces and only work in that context), are separate from the
reusable network_handshake.go functions.