Massive improvements to nonce handler
Defer the FullClose of the initiated close when we fail (rather than
leaving to the end). That had a chance of leaking if we fail anytime
during the process. Appropriate resets for the responder (as they didn't
open the connection). Remove max from nonce service (as we've killed
that field). Remove unnecessary pieces that caused panics, and
streamline some pieces (that help this handshake work).