Make `Threshold_nat` an integer.
In the process of submission and verification of tickets `W_k` we are
comparing them to the `Threshold_nat`. With our current definition of
`Threshold_nat` we will have a type mismatch cause `W_k` is a 256bits
int and `Threshold_nat` will be a (double) float. I would correct the
natural threshold definition with a floor function `Threshold_nat =
floor(N * Ticket_space / (Tokens_total / MINIMUM_STAKE))`. And same
applies to other definitions of thresholds.