Transform network public key into ETH address
We don't need to transfer NetworkPrivateKey to ethereum address but we
need to transfer NetworkPublicKey to ethereum address. During the
handshake, we don't have an access to the other party's private key.

I've also reworked method generating and transforming network keys to
return the public key as well. This helps with testing and may save us
from some type conversion madness.