change account generation to runtime and simplify Dockerfile
We now create the two necessary Ethereum accounts at runtime when
the docker container starts. We patch the genesis.json at runtime
to issue tokens to account1. We allocate mining rewards to account0.
Loggin output to STDOUT is a bit more structured now to clarify
boot phases of the container.