Moved received individual pub keys to function
Instead of keeping received individual public keys in a member's field
a function was created to get a slice of peers individual public keys
`A_j0` from public key share points `A_jk`.
We need to get zeroth element from the slice so we can use a function
for this, we won't duplicate a kept values this way.