Member group init func in tests expect dkg config
Utils functions in tests to initialize members group expects a dkg
configuration to be passed.
If the dkg configuration is not passed to the function the group will be
initialized with predefined dkg configuration from predefinedDKG()
function.
This is neccessary for some tests which requires custom `P` or `Q`
values to be set, e.g. small values for easier calculations.