Introduced ephemeral.KeyPair struct
The reason for introducing the `KeyPair` struct is to make the key
management easier for DKG group members. Each DKG group member needs to
hold a `KeyPair` of ephemeral keys that were used for creating a
`SymmetricKey` with each other group member. In case of an accusation,
accuser reveals the private ephemeral key that was used to create a
`SymmetricKey` with the accused party.