Expect SymmetricKey to be passed to PeerSharesMessage methods
S share and T share in PeerShareMessage needs to be encrypted. That's
why we expect SymmetricKey to be passed to initialize the message
(encryption) and that's why we expect SymmetricKey to be passed to
retrieve T or S share from the message (decryption).

Encryption/decryption mechanism has not been implemented yet. It will be
done separately.