Update CalculateEphemeralKeyPair
Rather than returning a list of public keys, we change membercore's
internal state to maintain a map of member id -> ephemeral key. This
allows us to directly store the generated ephemeral key pair for each
peer in the state, and broadcast it directly (rather than maintaining an
array and keeping a track of current index, and whether we've reached
our own index).