Added type alias representing member ID as uint32
Previously, the member ID was represented as int which has a
platform-dependent byte size. It's quite unwieldy for
marshalling/unmarshalling so we switch it to uint32 hidden under
MemberID alias. This way, any change of the underlying type will
be easier. We also know now what to expect in the serialization code.