Don't check for `nil` `ephemeralPublicKey` in the message
In phase 3, we expect to have symmetric keys generated with all other
group members who are not IA or DQ. If symmetric keys is not generated
for one or more group members, phase 3 yields an error.

A malicious actor can send a message with a missing
`ephemeralPublicKey`. With the not-nil check here, the message was
ignored and no symmetric key was generated. This was a recipe for a
failure in the next phase.

The way we should approach this problem is as follows:
When we receive a message in the glue layer we validate it - all
fields should be not nil. If we detect a nil field, we reject the
message and as a result mark the member as IA. We no longer need to
check `!= nil` in the protocol since it's redundant. We trust all the
passed messages are valid and they are addressed to us.