Generate only one ephemeral key pair message per group member
Instead of generating a separate ephemeral key pair message in
phase 1 for every other group member, we generate only one message
with all ephemeral public keys. The generated message contains
a map from the member ID to ephemeral public key generated
individually for that member.

The reason for that is to reduce the number of messages exchanged
in the group and to make the integration layer between the
network and GJKR protocol easier. All phase 1 messages need to
be broadcasted and stored in the evidence log by all group members
anyway.