Generate only one peer shares message per group member
Instead of generating a separate peer shares message in phase 3 for
every other group member, we generate only one message with all shares.
The generated message contains a map from the member ID to peer shares
generated individually for that member. Shares are encrypted with
symmetric key established with that member.

The reason for that is to reduce the number of messages exchanged in the
group and to make the integration layer between the network and GJKR
protocol easier. All phase 3 messages need to be broadcasted and stored
in the evidence log by all group members anyway.