Fix bug in encoding virtual staker index
Previously, the virtual staker value was little endian encoded into the
combined slice of bytes, which placed the vs_index to the front of the
array. Rather, we little endian encode the value into an intermediary
slice of bytes, and then append this to the end of the combined bytes.