Filter out staker added event by staker ID
When random beacon is initialized, the node is registered as staker.
There was a bug in the watching code - we were reacting on any
OnStakerAdded event and continuing with staker's ID from that event.
It was often happening, that some other staker was registered in a
meantime and instead of waiting for the current node to be added as
staker, we were continuing. Also, what's even worse, we use the staker's
ID from OnStakerAdded event to update current node's ID after the
OnStakerAdded event happens. It was often happening that we were
setting current node's ID to some other node's ID.