change networkid for local kubernetes ethereum network and rename scripts
1. The NETOWORKID parameter in the yaml files needed to be more unique
   and not just a ISO date value. Changed to a ten digit random number.
2. Renamed the start and shutdown scripts for the local Kubernetes
   cluster to `up` and `down` to save some typing.

expose tx-node on localhost:8545 via type LoadBalancer

To simplify testing with the Ethereum test cluster we now expose
the geth RPC port 8545 of the tx-node on localhost.

add gratuitous output of dashboard URL when booting the Ethereum cluster

add diagram showing interactions between Ethereum testnet components

add ignoring of keystore directory

add hostvolume support for DATADIR to save chain data

We want to be able to persist existing chain data in case pods
are recycled by Kubernetes. To do this we introduce a HOSTVOLUME
parameter that if set will be used to store chain data when geth
boots.

make number of Keep client accounts created a Docker build parameter

Depending on the number of Keep clients we want to run concurrently
attached to the Ethereum testnet we need to pregenerate a matching
number of accounts and prefund them. The Docker build command now
takes an integer parameter KEEP_ACCOUNTS that specifies the number
of accounts to create. Accounts are created and added to the genesis.json
file where they are then prefunded.

set ethash.dagdir to equal DATADIR when booting geth

We need to store the ethash DAG in the same location as other chain data
so we can manage all geth data on a persistent volume.

add support for parameterized directories in hostPath persistent volumes

To persist chain data on a local volume when running on Docker Desktop
or in the cloud we need to create sub-directories for each pod
instance in case we are writing to a shared volume.
We need to inform the bootNodeReporter about the changed path to
the IPC file that gives it access to geth via IPC.

Comma delimit each eth account object so that we have valid json

add last minute fixes to generate accounts and genesis.json etc.

- We now use a locally mounted volume for chaindata storage. Defaults
  are provided.
- We now create distinct sub-directories for each pod if the `DESKTOP`
  switch is true.
- The address for the mining Ethereum account is now stored in a file
  named `mining_account`.
- The test for an existing chaindata directory used the wrong path and
  we fixed that. We only initialize the chain if the directory doesn't
  exist.
- We now set the ETH_IPC_PATH explicitly and this is also reflected in
  the yaml files for the miner and tx pods. Defaults are provided.
- We are now using all predefined environment values when booting
  `geth`.
- We now use `jq` and a template to generate the genesis.json file.
  This way we ensure correct JSON syntax at Docker build time.
- Added `keep_accounts` to .gitignore

fix typo in filename for genesis-template.json

Add arg to send jq input to stdout

add two more backslashes to escape the quotes to make balance a string

The bash shell was eliminating the quotes around the decimal
defining the account balance. We needed to add a backslash in
front of each quote to turn the balance into a string.

fix up paths and fix quotes in `geth` parameters

Some of the parameterized `geth` arguments lacked proper quotation
marks because the shell was eating them. They are now properly
quoted with single ticks.

simplify PersistentVolume usage when running via Docker Desktop

We want to be able to reuse chain data by reusing the same local volume
for multiple restarts of the local Ethereum cluster. To keep this
simple we define one volume for the miner node and one volume for
the tx node. If the user needs to run more than one replica of miner
or tx node in the local Ethereum network, she needs to undefine
`HOSTVOLUME` and switch to ephemeral storage.

Remove quotes so RPCAPI args are taken in

Remove single quotes from wsorigins so we actually accept ws connections

fix ETH_IPC_PATH_DEFAULT and add `geth.ipc` at the end

fix issue with RPCAPI service list quoting

Indent one tab

Revert "Indent one tab"

This reverts commit cdd1b86951ee9589450156e8a19106c4d79373ea.

Escape quotes on wsorigins so the env var doesn't get set to a list of the directory contents

Unescape qutoes for wsorigin so that we get an unquoted star for the wsorigins value

move ws args back to the geth command

add more severe warning about handling passwords when generation images

We are using a static password sourced from a file to pregenerate
Ethereum accounts. Hopefully the added warning in the README will
caution users when generating geth-node images to only use this
setup in a private testnet environment.

add first draft of script to extract keyfiles and create ConfigMap

When baking a geth-node image for the Ethereum test net we pre-generate
Ethereum accounts for Keep clients. This script extracts the keyfiles
and creates a Kubernetes ConfigMap that maps each account ID to the
keyfile. Kubernetes will later query this ConfigMap to fetch the
appropriate keyfile for each Keep client node.

fix minor bug with argument processing in manage-keyfiles-as-ConfigMaps

We were shifting twice instead of once. Fixed.

improve key naming to store keyfiles for Keep clients in ConfigMaps

instead of just the account ID we now use the pattern
<keystore-accountID> as keys to store the keyfiles in a ConfigMap.

update README and add draft of keystore-configmap-job

- The README for geth-dev needed a section that explains how to check
  the version of geth deployed in a generated Docker image.
- We add an initial draft of a Kubernetes job that uses a configMap
  to manage keystore keys for Ethereum accounts (still buggy)